requirements.txt
python-telegram-bot==12.7
python-dotenv==1.0.1